os: KeyValueDB: Add virtual raw_key() function to return (prefix,key) pair
In certain cases, such as when we are iterating over the whole key space,
it is useful to obtain the (prefix, key) pair. We could easily obtain the
key as is, but given that each implementation may deal with prefixes and
keys differently (for instance, the LevelDBStore will use '\0' as the
separator for the 'prefix' and 'key' components), we would rather have a
function responsible for returning a pair instead.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>