os: LevelDBStore: implement the KeyValueDB's raw_key() function
If we were to use solely the key() function, whenever we had a key with,
say, prefix 'Foo' and key 'Bar', the key() function would return something
similar to 'FooBar'. Since the separator used on the LevelDBStore is '\0',
but considering that different implementations of KeyValueDB may use
different separators, anyone using a K/V implementation would have to be
aware of the separator and parse the prefix and the key if they so wanted
to do it during an iteration over the whole store's key space.

As such, implement the raw_key() function which will be responsible for
returning a pair (prefix,key), leaving the parsing and all the
implementation-specific details for the implementation itself.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>