mon: MonitorDBStore: add clear() function to remove all keys from the store
We also provide a finer-grained function that should receive a set of
strings containing exceptions. Basically, any key in the store with the
same prefix as any value in this exceptions set will not be removed.

This is useful to clear the Paxos' and the services' state from the store
without messing the Monitor-specific keys, which if removed would make the
monitor unusable.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>