mon: Monitor: Add initial support for synchronization (WARNING: WIP)
This is very on-going work, and I've actually stopped working on it for the
time being while addressing other pieces of this puzzle. And I feel I can't
emphasize this enough: this is on-going work.

There is a lot of stuff missing, some other stuff was completely left in
a limbo of half-done/half-undone, and it will certainly NOT WORK.

Also, there are a couple of things that are to be replaced, some other
stuff will be stripped away.

Please, do not waste time reviewing it.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>