mon: MonitorDBStore: Use LevelDBStore's thread-safe iterator
We need to read the store while still performing updates due to Paxos
proposals. Therefore we need to make sure we always see a consistent state
despite what may be happening. By using a thread-safe iterator from
LevelDBStore, we guarantee that we do not incur in unforeseen problems
regarding keys being added or removed while iterating over the available
keys in the store.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>