os: KeyValueDB: create whole-space iterator and snapshot iterator
In-a-nutshell-version:
  1) Create a whole-space iterator interface, and implement the already
     existing, prefix-based iterator in terms of the new whole-space
     iterator;
  2) Add support for snapshot iterators.

This patch introduces a significant change on the architecture of
KeyValueDB's iterator, although its interface remains the same.

Before this patch, the KeyValueDB simply defined an interface for a
prefix-based interface, to be implemented by derivative classes. Being
constrained by a prefix-based approach to iterate over the store only makes
sense when we know which prefixes we want to iterate over, but for that we
must know about the prefixes beforehand. This approach didn't work when one
wanted to iterate over the whole key space, without any previous awareness
about the keys and their prefixes.

This patch introduces a new interface for a whole-space iterator, to be
implemented by derivative classes, which is prefix-independent. We also
define an abstract function to obtain this iterator, which must also be
implemented by the derivative class. With this interface in place, we are
then able to implement a prefix-dependent iterator in terms of the
whole-space iterator, which will be offered by the KeyValueDB class itself.

Furthermore, we also introduce support for a snapshot iterator. The task
of implementing such iterator is delegated to any derivative class.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>