os: LevelDBStore: Implement whole-space and snapshot iterators
Following the changes to KeyValueDB, we get rid of the previous,
prefix-dependent iterator implementation, and we create a whole-space
iterator for a leveldb k/v store. We also add a snapshot iterator that
relies on the said whole-space iterator.

This snapshot iterator relies on leveldb's snapshots, and is intended as
a thread-safe iterator for those occasions when preventing concurrent
modification is not an option.

For instance, we need to iterate over a known, consistent store state
during monitor synchronization, while allowing updates to be performed for
Paxos sake -- and that means both insertion and removal of keys, which
could be, at best, problematic or simply bear unforeseen issues that could
ultimately only be seen during full-fledged concurrent accesses to the
store.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>