test: cli: mon-store-tool: Add Validator class to validate two stores
This class will perform two kinds of verification: strict and relaxed.

A strict verification will check if all keys and values on both stores
match and iterating them at the same time on either store will return the
same keys and values. I.e., the first key on store A is the same as the
first key on store B, and both hold the same value; etc.

A relaxed verification will check if all the keys existing on A also exist
on B, and vice-versa, outputing any mismatches regarding values or
non-existing keys on either of the stores.

For instance, say that store A has the following keys/values:
  (foo,bar) = 'foobar'
  (boo,yah) = 'booyah'
  (boo,gah) = 'boogah'

and store B has the following keys/values:
  (bar,foo) = 'barfoo'
  (boo,yah) = 'booyah'
  (boo,gah) = 'gahboo'

A Relaxed verification will endup with the following results:

  key (boo,yah) value 'booyah' matches on both stores
  key (foo,bar) exists on A but does not exist on B
  key (bar,foo) exists on B but does not exist on A
  key (boo,gah) value mismatch on A ('boogah') and B ('gahboo')

This will be useful to check on two different stores during Monitor sync.

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>