filejournal: add a min time interval between writes
If the journal devices has NVRAM, it may end up giving us near-zero latency
for small writes, which means we will write each event individually.  For
a stream of small writes, this can be suboptimal, as it will turn 4k writes
into 12k entries, and end up saturating the disk 3x sooner.

This patch will force us to wait some minimum amount of time between
dispatching a new write to the journal device.  The tradeoff is a slightly
higher latency in exchange for higher throughput.

Signed-off-by: Sage Weil <sage@inktank.com>